requirements.txt
streamlit
pandas
requests
python-jose[cryptography]
python-dotenv
